+++ sdks/python/apache_beam/runners/pipeline_context_test.py
@@ -53,12 +53,12 @@ class PipelineContextTest(unittest.TestCase):
     self.assertEqual(
         coders.BytesCoder(), context2.coders.get_by_id(bytes_coder_ref))
 
-  def test_id_assignment_spans_multiple_instances(self):
+  def test_common_id_assignment(self):
     context = pipeline_context.PipelineContext()
     float_coder_ref = context.coders.get_id(coders.FloatCoder())
     bytes_coder_ref = context.coders.get_id(coders.BytesCoder())
     context2 = pipeline_context.PipelineContext(
-        component_id_context=context.component_id_context)
+        component_id_map=context.component_id_map)
 
     bytes_coder_ref2 = context2.coders.get_id(coders.BytesCoder())
     float_coder_ref2 = context2.coders.get_id(coders.FloatCoder())
